cannot delete post if user didn't create it

+++ server/controllers/feed.js
@@ -128,7 +128,7 @@ exports.updatePost = (req, res, next) => {
 
       if (post.creator.toString() !== req.userId) {
         const error = new Error(
-          'cannot delete ' + postId + ' as not authorized'
+          'cannot update ' + postId + ' as not authorized'
         );
         error.statusCode = 403;
         throw error;
@@ -162,7 +162,13 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
-
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error(
+          'cannot delete ' + postId + ' as not authorized'
+        );
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
